# web/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MessagesIn": {
        "properties": {
          "messages": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Messages",
            "type": "array"
          }
        },
        "required": [
          "messages"
        ],
        "title": "MessagesIn",
        "type": "object"
      },
      "PersonaIn": {
        "properties": {
          "dept": {
            "default": "",
            "title": "Dept",
            "type": "string"
          },
          "interest_keywords": {
            "items": {
              "type": "string"
            },
            "title": "Interest Keywords",
            "type": "array"
          },
          "interest_lv3": {
            "items": {
              "type": "string"
            },
            "title": "Interest Lv3",
            "type": "array"
          },
          "interest_tasks": {
            "items": {
              "type": "string"
            },
            "title": "Interest Tasks",
            "type": "array"
          },
          "job": {
            "default": "",
            "title": "Job",
            "type": "string"
          },
          "muted_keywords": {
            "items": {
              "type": "string"
            },
            "title": "Muted Keywords",
            "type": "array"
          },
          "name": {
            "default": "",
            "title": "Name",
            "type": "string"
          },
          "team": {
            "default": "",
            "title": "Team",
            "type": "string"
          }
        },
        "title": "PersonaIn",
        "type": "object"
      },
      "PersonaModel": {
        "properties": {
          "dept": {
            "default": "",
            "title": "Dept",
            "type": "string"
          },
          "derived_at": {
            "default": "",
            "title": "Derived At",
            "type": "string"
          },
          "derived_interests": {
            "items": {
              "type": "string"
            },
            "title": "Derived Interests",
            "type": "array"
          },
          "derived_source": {
            "default": "",
            "title": "Derived Source",
            "type": "string"
          },
          "interest_keywords": {
            "items": {
              "type": "string"
            },
            "title": "Interest Keywords",
            "type": "array"
          },
          "interest_lv3": {
            "items": {
              "type": "string"
            },
            "title": "Interest Lv3",
            "type": "array"
          },
          "interest_tasks": {
            "items": {
              "type": "string"
            },
            "title": "Interest Tasks",
            "type": "array"
          },
          "is_set": {
            "default": false,
            "title": "Is Set",
            "type": "boolean"
          },
          "job": {
            "default": "",
            "title": "Job",
            "type": "string"
          },
          "label": {
            "default": "",
            "title": "Label",
            "type": "string"
          },
          "matched_processes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Matched Processes",
            "type": "array"
          },
          "muted_keywords": {
            "items": {
              "type": "string"
            },
            "title": "Muted Keywords",
            "type": "array"
          },
          "name": {
            "default": "",
            "title": "Name",
            "type": "string"
          },
          "team": {
            "default": "",
            "title": "Team",
            "type": "string"
          }
        },
        "title": "PersonaModel",
        "type": "object"
      },
...
      "ThreadCreateIn": {
        "properties": {
          "first_message": {
            "default": "",
            "title": "First Message",
            "type": "string"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          }
        },
        "title": "ThreadCreateIn",
        "type": "object"
      },
      "ThreadUpdateIn": {
        "properties": {
          "pinned": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pinned"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "ThreadUpdateIn",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "조선소 작업정의 × 외부 기술동향 매칭 — React 전환용 백엔드 계약",
    "title": "InsightBoard API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/persona": {
      "get": {
        "operationId": "get_persona_api_persona_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Workspace-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonaModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Persona",
        "tags": [
          "persona"
        ]
      },
      "put": {
        "operationId": "put_persona_api_persona_put",
        "parameters": [
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Workspace-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonaIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonaModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Persona",
        "tags": [
          "persona"
        ]
      }
    },
    "/api/persona/derive": {
      "post": {
        "operationId": "derive_persona_api_persona_derive_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Workspace-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonaModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Derive Persona",
        "tags": [
          "persona"
        ]
      }
    },
    "/api/persona/reset": {
      "post": {
        "operationId": "reset_persona_api_persona_reset_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Workspace-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonaModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reset Persona",
        "tags": [
          "persona"
        ]
      }
    },
...
    "/api/threads": {
      "get": {
        "operationId": "list_threads_api_threads_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Workspace-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Threads Api Threads Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Threads",
        "tags": [
          "threads"
        ]
      },
      "post": {
        "description": "스레드 생성. title 이 비고 first_message 가 있으면 LLM 으로 제목 자동 생성.\n\n과거엔 프런트가 첫 메시지를 36자로 자른 제목을 보냈다 → 의미 없는 제목.\n`sola.thread_title.generate` 는 캐시 + 룰 fallback 내장이라 LLM 미설정·실패에도\n안전(예외 없음).",
        "operationId": "create_thread_api_threads_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Workspace-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThreadCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Thread Api Threads Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Thread",
        "tags": [
          "threads"
        ]
      }
    },
    "/api/threads/{thread_id}": {
      "delete": {
        "operationId": "delete_thread_api_threads__thread_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "thread_id",
            "required": true,
            "schema": {
              "title": "Thread Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Thread Api Threads  Thread Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Thread",
        "tags": [
          "threads"
        ]
      },
      "get": {
        "operationId": "get_thread_api_threads__thread_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "thread_id",
            "required": true,
            "schema": {
              "title": "Thread Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Thread Api Threads  Thread Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Thread",
        "tags": [
          "threads"
        ]
      },
      "patch": {
        "operationId": "update_thread_api_threads__thread_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "thread_id",
            "required": true,
            "schema": {
              "title": "Thread Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThreadUpdateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Update Thread Api Threads  Thread Id  Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Thread",
        "tags": [
          "threads"
        ]
      }
    },
    "/api/threads/{thread_id}/messages": {
      "get": {
        "operationId": "get_messages_api_threads__thread_id__messages_get",
        "parameters": [
          {
            "in": "path",
            "name": "thread_id",
            "required": true,
            "schema": {
              "title": "Thread Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Get Messages Api Threads  Thread Id  Messages Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Messages",
        "tags": [
          "threads"
        ]
      },
      "put": {
        "operationId": "put_messages_api_threads__thread_id__messages_put",
        "parameters": [
          {
            "in": "path",
            "name": "thread_id",
            "required": true,
            "schema": {
              "title": "Thread Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessagesIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Put Messages Api Threads  Thread Id  Messages Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Messages",
        "tags": [
          "threads"
        ]
      }
    },
...
  }
}